Update csharp and java for Windows
git-svn-id: https://code.crt.realtors.org/svn/librets/librets/trunk@1564 fe7ac059-47f5-0310-8599-d0339b4e125c

+++ project/swig/csharp/CppInputStream.cs
@@ -64,7 +64,7 @@ public class CppInputStream : Stream
 
     public override int Read([In,Out] byte[] buffer, int offset, int count)
     {
-         return mBridge.read(buffer, offset, count);
+         return (int) mBridge.read(buffer, (uint) offset, (uint) count);
     }
     
     private InputStreamBridge mBridge;
